add admin menu

+++ user_control/delete.php
@@ -1,15 +1,23 @@
 <?php
 require_once($_SERVER['DOCUMENT_ROOT'] . '/e-commerce-site/db_connect.php');
+require_once("../utils/user-auth.php");
 $conn = new Database();
+
+$userAuth = new userAuth($conn);
+$isDelete = $userAuth->checkDeletePermission("5");
+
+if (!$isDelete) {
+  header("Location: ../pages/user.php");
+  exit();
+}
+
 $id = $_GET["id"];
 $sql = "DELETE FROM user WHERE id=$id";
 
 $result = $conn->query($sql);
 
 if ($result) {
-  session_start();
-  $_SESSION["product_msg"] = "xóa sản phẩm thành công";
-  header("Location: ../user.php");
+  header("Location: ../pages/user.php");
 }
 
 $conn->close();
